requirements.txt
# FastAPI und Web-Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Datenbank
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Authentifizierung
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validierung und Konfiguration
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utilities
python-multipart>=0.0.6
email-validator>=2.1.0
